fix: run integration tests on large runners
This is something I'd been going back and forth on, but now
we've got a reason to use it. With all the large dependencies
in the developer framework, we are eclipsing the relatively
small amount of stroage space on GitHub's standard runners.
We already have a pool of larger runners being used to build
our docker images faster, but they also have more disk space.
If we notice jobs being queued, we can always add more large
runners to the pool

Signed-off-by: Cullen Taylor <cullen@eggshell.me>